refactor: extract components from examples
