+++ example/lib/examples/basic_demo.dart
@@ -1,0 +1,90 @@
+import 'package:example/widgets/bottom_buttons_row.dart';
+import 'package:example/widgets/card_overlay.dart';
+import 'package:example/widgets/example_card.dart';
+import 'package:flutter/material.dart';
+import 'package:swipable_stack/swipable_stack.dart';
+
+const _images = [
+  'images/image_5.jpg',
+  'images/image_3.jpg',
+  'images/image_4.jpg',
+];
+
+class BasicDemo extends StatefulWidget {
+  const BasicDemo({Key? key}) : super(key: key);
+
+  @override
+  _BasicDemoState createState() => _BasicDemoState();
+}
+
+class _BasicDemoState extends State<BasicDemo> {
+  late final SwipableStackController _controller;
+
+  void _listenController() {
+    setState(() {});
+  }
+
+  @override
+  void initState() {
+    super.initState();
+    _controller = SwipableStackController()..addListener(_listenController);
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    _controller.removeListener(_listenController);
+    _controller.dispose();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: SafeArea(
+        child: Stack(
+          children: [
+            Positioned.fill(
+              child: Padding(
+                padding: const EdgeInsets.all(8.0),
+                child: SwipableStack(
+                  controller: _controller,
+                  stackClipBehaviour: Clip.none,
+                  onSwipeCompleted: (index, direction) {
+                    print('$index, $direction');
+                  },
+                  horizontalSwipeThreshold: 0.8,
+                  verticalSwipeThreshold: 0.8,
+                  overlayBuilder: (
+                    context,
+                    constraints,
+                    index,
+                    direction,
+                    swipeProgress,
+                  ) =>
+                      CardOverlay(
+                    swipeProgress: swipeProgress,
+                    direction: direction,
+                  ),
+                  builder: (context, index, constraints) {
+                    final itemIndex = index % _images.length;
+                    return ExampleCard(
+                      name: 'Sample No.${itemIndex + 1}',
+                      assetPath: _images[itemIndex],
+                    );
+                  },
+                ),
+              ),
+            ),
+            BottomButtonsRow(
+              onSwipe: (direction) {
+                _controller.next(swipeDirection: direction);
+              },
+              onRewindTap: _controller.rewind,
+              canRewind: _controller.canRewind,
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
